Make a clear separation for api filtering interface and db-based impl
 * refactor to separate
 * define custom type class implementation for operations and fees
 * revise the access modifiers for database filtering operations to be
   more strict
